fixes after PR
